Disable Discord, update README, begin refactor
