+++ config/auth.json.example
