+++ lib/doorChat.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const chalk = require('chalk');
 const path = require('path');
 const commandDirectory = '/plugins';
@@ -13,9 +15,11 @@ module.exports = Doorman => {
       console.log(err);
     }
   };
+
   Doorman.commandCount = () => {
     return Object.keys(Doorman.Commands).length;
   };
+
   Doorman.setupCommands = () => {
     Doorman.Commands = {};
 
@@ -41,6 +45,8 @@ module.exports = Doorman => {
     // Load simple commands from json file
     let jsonCommands = [];
     try {
+      // TODO: refactor how external data is loaded
+      // I/O should not take place inside of the module
       jsonCommands = Doorman.getJsonObject('/config/commands.json');
     } catch (err) { }
 
@@ -84,10 +90,11 @@ module.exports = Doorman => {
       });
     }
   };
+
   Doorman.setupCommands();
 
   // Discord
-  require('./discord/discordSupport')(Doorman);
+  //require('./discord/discordSupport')(Doorman);
 
   // Slack
   require('./slack/slackSupport')(Doorman);
