+++ lib/doorMaki.js
@@ -1,9 +1,11 @@
+'use strict';
+
 const Maki = require('maki');
-const makiConfig = require('../config/makiConfig');
+const config = require('../config/maki');
 
 module.exports = Doorman => {
   // delicious maki...
-  Doorman.maki = new Maki(makiConfig);
+  Doorman.maki = new Maki(config);
   Doorman.maki.use(require('maki-client-level'));
   Doorman.maki.start();
 };
