+++ lib/slack/onEvent/auth-and-connect.js
@@ -1,8 +1,9 @@
 module.exports = function (Doorman) {
-  Doorman.Slack.on(Doorman.Slack.c_events.RTM.AUTHENTICATED, rtmStartData => {
-    console.log(`Logged into Slack! Name:${rtmStartData.self.name}, Team:${rtmStartData.team.name}`);
+  Doorman.Slack.on('ready', rtmStartData => {
+    //console.log(`Logged into Slack! Name: ${rtmStartData.self.name}, Team:${rtmStartData.team.name}`);
+    console.log(`Logged into Slack!`);
   });
-  Doorman.Slack.on(Doorman.Slack.c_events.RTM.RTM_CONNECTION_OPENED, () => {
+  Doorman.Slack.on('connected', () => {
     console.log(`Connection to Slack Successful!`);
   });
 };
