+++ lib/slack/onEvent/message.js
@@ -79,28 +79,33 @@ module.exports = function (Doorman) {
   }
 
   function checkMessageForCommand (msg, isEdit) {
-    const self = Doorman.Slack.dataStore.getBotByName('Doorman');
+    // TODO: re-bind Doorman.Config to Doorman.config
+    console.log('msg:', msg);
+
+    //const bot = Doorman.Slack.dataStore.getBotByName(Doorman.Config.name);
     msg.author = `<@${msg.user}>`;
     // Drop our own messages to prevent feedback loops
-    if (msg.user === self.id) {
+    if (msg.subtype && msg.subtype === 'bot_message') {
       return;
     }
     if (Doorman.Config.debug) {
       console.log('message received:', msg.type, msg.subtype, 'interpreting...');
     }
+
     // Check for mention
-    if (msg.text.split(' ')[0] === `<@${self.id}>`) {
+    /* if (msg.text.split(' ')[0] === `<@${bot.id}>`) {
       if (Doorman.Config.elizaEnabled) {
         // If Eliza AI is enabled, respond to @mention
-        const message = msg.text.replace(`<@${self.id}> `, '');
+        const message = msg.text.replace(`<@${bot.id}> `, '');
         Doorman.Slack.sendMessage(Doorman.Eliza.transform(message), msg.channel);
         return;
       }
       Doorman.Slack.sendMessage('Yes?', msg.channel);
       return;
-    }
+    }*/
+
     // Check for IM
-    if (Doorman.Slack.dataStore.getDMById(msg.channel)) {
+    /*if (Doorman.Slack.dataStore.getDMById(msg.channel)) {
       if (msg.text.startsWith(Doorman.Config.commandPrefix) && msg.text.split(' ')[0].substring(Doorman.Config.commandPrefix.length).toLowerCase() === 'reload') {
         Doorman.setupCommands();
         Doorman.setupSlackCommands();
@@ -110,7 +115,8 @@ module.exports = function (Doorman) {
       }
       Doorman.Slack.sendMessage(`I don't respond to direct messages.`, msg.channel);
       return;
-    }
+    }*/
+
     // Check if message is a command
     if (msg.text.startsWith(Doorman.Config.commandPrefix)) {
       const allCommands = Object.assign(Doorman.Commands, Doorman.slackCommands);
@@ -193,5 +199,5 @@ module.exports = function (Doorman) {
     }
   }
 
-  Doorman.Slack.on(Doorman.Slack.rtm_events.MESSAGE, msg => checkMessageForCommand(msg, false));
+  Doorman.Slack.on('message', msg => checkMessageForCommand(msg, false));
 };
