+++ lib/slack/onEvent/team-join.js
@@ -1,5 +1,5 @@
 module.exports = function (Doorman) {
-  Doorman.Slack.on(Doorman.Slack.rtm_events.TEAM_JOIN, event => {
+  Doorman.Slack.on('team_join', event => {
     Doorman.Slack.sendMessage(`@here, please Welcome ${event.user.name} to ${Doorman.Config.slack.teamName}!`, Doorman.Slack.dataStore.getChannelByName(Doorman.Config.slack.welcomeChannel).id);
   });
 };
