+++ lib/slack/slackSupport.js
@@ -1,11 +1,13 @@
+'use strict';
+
 const chalk = require('chalk');
-const RtmClient = require('@slack/client').RtmClient;
+const RTMClient = require('@slack/client').RTMClient;
 
 module.exports = Doorman => {
   console.log(chalk.magenta(`Slack Enabled... Starting.`));
   if (Doorman.Auth.slack.bot_token) {
     console.log('Logging in to Slack...');
-    Doorman.Slack = new RtmClient(Doorman.Auth.slack.bot_token);
+    Doorman.Slack = new RTMClient(Doorman.Auth.slack.bot_token);
     Doorman.Slack.c_events = require('@slack/client').CLIENT_EVENTS;
     Doorman.Slack.rtm_events = require('@slack/client').RTM_EVENTS;
     Doorman.Slack.start();
