+++ test/doorman.unit.js
@@ -7,4 +7,13 @@ describe('Doorman', function () {
   it('should expose a constructor', function () {
     assert(typeof Doorman, 'function');
   });
+
+  it('can handle a message', function (done) {
+    let doorman = new Doorman();
+    
+    doorman.on('input', function (data) {
+      assert.ok(data);
+      done();
+    });
+  });
 });
